% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemosensorsClass.R, R/affinity.lib.R, R/methods.R
\docType{methods}
\name{compute-methods}
\alias{compute-methods}
\alias{computeAffinity,ANY-method}
\title{Method computeAffinity.}
\usage{
computeAffinity(object, method, norm = "none", ...)

\S4method{computeAffinity}{ANY}(object, method, norm = "none", ...)
}
\description{
Method computeAffinity.

Compute methods in R package chemosensors.
}

