\name{ConcNoiseModel}
\alias{ConcNoiseModel}
\title{Class ConcNoiseModel.}
\description{
  Class \code{\link{ConcNoiseModel}} generates a
  concentration noise.
}
\details{
  Slots of the class: \tabular{rl}{ \code{csd} \tab
  Parameter of standard deviation used to generate the
  noise.  The deault value is 0.1. \cr \code{cntype} \tab
  Noise type. One type is implemented, 'logconc'. \cr
  \code{cnlogf} \tab Scaling factor for log-terms used to
  generate the noise.  The default value is c(1, 1, 2). \cr
  }
}
\keyword{ConcNoiseModel-class}

