\name{DriftCommonModel-class}
\alias{DriftCommonModel}
\alias{DriftCommonModel-class}
\title{Class DriftCommonModel.}
\usage{
  DriftCommonModel(dmodel, ...)
}
\arguments{
  \item{dmodel}{Name of the drift model. The only supported
  value is \code{"cpc"}.}

  \item{...}{Parameters of constructor.}
}
\value{
  AN object of class \code{DriftCommonModel}
}
\description{
  Class DriftCommonModel.

  Constructor method of DriftCommonModel Class.
}

