% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemosensorsClass.R, R/ChemosensorsClassMethods.R, R/ConcNoiseModelClass.R, R/DriftNoiseModelClass.R, R/ScenarioClass.R, R/SorptionModelClass.R, R/SensorDynamicsClass.R, R/SensorModelClass.R, R/SensorNoiseModelClass.R, R/SensorArrayClass.R, R/SensorArrayClassMethods.R, R/SensorDynamicsClassMethods.R, R/SensorModelClassMethods.R, R/SensorNoiseModelClassMethods.R, R/methods.R
\docType{methods}
\name{get-methods}
\alias{alpha,SorptionModel-method}
\alias{beta,SensorModel-method}
\alias{coef,ANY-method}
\alias{coefficients,SensorArray-method}
\alias{coefficients,SensorDynamics-method}
\alias{coefficients,SensorModel-method}
\alias{coefnames,SensorNoiseModel-method}
\alias{concUnits,ANY-method}
\alias{concUnitsInt,ANY-method}
\alias{enableDyn,SensorDynamics-method}
\alias{enableSorption,SensorArray-method}
\alias{gases,ANY-method}
\alias{gases,missing-method}
\alias{get-methods}
\alias{gind,ANY-method}
\alias{gind,missing-method}
\alias{gnames,ANY-method}
\alias{gnames,missing-method}
\alias{idx,ANY-method}
\alias{modelName,DriftNoiseModel-method}
\alias{modelName,SensorModel-method}
\alias{ncoef,SensorDynamics-method}
\alias{ncoef,SensorModel-method}
\alias{ncoef,SensorNoiseModel-method}
\alias{ngases,ANY-method}
\alias{ngases,missing-method}
\alias{nsensors,ANY-method}
\alias{nsensors,SorptionModel-method}
\alias{num,ANY-method}
\alias{snames,ANY-method}
\alias{tunit,Scenario-method}
\alias{tunit,SensorDynamics-method}
\alias{type,ConcNoiseModel-method}
\alias{type,DriftNoiseModel-method}
\alias{type,SensorNoiseModel-method}
\title{Method alpha.}
\usage{
alpha(x)

beta(x)

modelName(x)

tunit(x)

enableSorption(x)

enableDyn(x)

num(x)

idx(x)

gases(x)

gind(x)

ngases(x)

gnames(x)

coefficients(object, ...)

coef(object, ...)

nsensors(x)

snames(x, ...)

concUnits(x)

concUnitsInt(x)

type(x)

coefnames(x)

ncoef(x)

\S4method{num}{ANY}(x)

\S4method{idx}{ANY}(x)

\S4method{gases}{ANY}(x)

\S4method{gind}{ANY}(x)

\S4method{ngases}{ANY}(x)

\S4method{gnames}{ANY}(x)

\S4method{gases}{missing}(x)

\S4method{gind}{missing}(x)

\S4method{ngases}{missing}(x)

\S4method{gnames}{missing}(x)

\S4method{nsensors}{ANY}(x)

\S4method{snames}{ANY}(x, sensor.names = "short", ...)

\S4method{concUnits}{ANY}(x)

\S4method{concUnitsInt}{ANY}(x)

\S4method{coef}{ANY}(object, ...)

\S4method{type}{ConcNoiseModel}(x)

\S4method{modelName}{DriftNoiseModel}(x)

\S4method{type}{DriftNoiseModel}(x)

\S4method{tunit}{Scenario}(x)

\S4method{alpha}{SorptionModel}(x)

\S4method{nsensors}{SorptionModel}(x)

\S4method{tunit}{SensorDynamics}(x)

\S4method{enableDyn}{SensorDynamics}(x)

\S4method{modelName}{SensorModel}(x)

\S4method{beta}{SensorModel}(x)

\S4method{type}{SensorNoiseModel}(x)

\S4method{enableSorption}{SensorArray}(x)

\S4method{coefficients}{SensorArray}(object, type, ...)

\S4method{coefficients}{SensorDynamics}(object, ...)

\S4method{ncoef}{SensorDynamics}(x)

\S4method{coefficients}{SensorModel}(object, ...)

\S4method{ncoef}{SensorModel}(x)

\S4method{ncoef}{SensorNoiseModel}(x)

\S4method{coefnames}{SensorNoiseModel}(x)
}
\description{
Method alpha.

Method beta.

Method modelName.

Method tunit.

Method enableSorption.

Method enableDyn.

Method num.

Method idx.

Method gases.

Method gind.

Method ngases.

Method gnames.

Method coefficients.

Method coef.

Method nsensors.

Method snames.

Method concUnits.

Method concUnitsInt.

Method type.

Method coefnames.

Method ncoef.

Get Methods in R package chemosensors.
}

