% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemosensorsClass.R, R/ChemosensorsClassMethods.R, R/ConcNoiseModelClassMethods.R, R/DriftNoiseModelClassMethods.R, R/ScenarioClassMethods.R, R/SensorArrayClassMethods.R, R/SensorDynamicsClassMethods.R, R/SensorModelClassMethods.R, R/SensorNoiseModelClassMethods.R, R/SorptionModelClassMethods.R, R/methods.R
\docType{methods}
\name{model-methods}
\alias{concMax,ANY-method}
\alias{concMin,ANY-method}
\alias{concSample,ANY-method}
\alias{concSample,Scenario-method}
\alias{concSampleDyn,ANY-method}
\alias{getConc,Scenario-method}
\alias{model-methods}
\alias{predict,ConcNoiseModel-method}
\alias{predict,DriftNoiseModel-method}
\alias{predict,SensorArray-method}
\alias{predict,SensorDynamics-method}
\alias{predict,SensorModel-method}
\alias{predict,SensorNoiseModel-method}
\alias{predict,SorptionModel-method}
\title{Method concMin.}
\usage{
concMin(object, ...)

concMax(object, ...)

concSample(object, ...)

concSampleDyn(object, ...)

predict(object, ...)

getConc(object, ...)

\S4method{concMin}{ANY}(object, concUnits = object@concUnits, ...)

\S4method{concMax}{ANY}(object, concUnits = object@concUnits, ...)

\S4method{concSample}{ANY}(object, type, n, gases = 0,
  concUnits = "default", ...)

\S4method{concSampleDyn}{ANY}(object, type, n, gases = 0,
  concUnits = "default", ...)

\S4method{predict}{ConcNoiseModel}(object, conc, concUnits = "default", ...)

\S4method{predict}{DriftNoiseModel}(object, sdata, ...)

\S4method{concSample}{Scenario}(object, n = 1, ...)

\S4method{getConc}{Scenario}(object, set, cf, ...)

\S4method{predict}{SensorArray}(object, conc, coef = "numeric",
  concUnits = "default", ...)

\S4method{predict}{SensorDynamics}(object, conc, sdata, sensors, ...)

\S4method{predict}{SensorModel}(object, conc, coef = "numeric",
  concUnits = "default", ...)

\S4method{predict}{SensorNoiseModel}(object, coef, n, ...)

\S4method{predict}{SorptionModel}(object, conc, concUnits = "default", ...)
}
\description{
Method concMin.

Method concMax.

Method concSample.

Method concSampleDyn.

Method predict.

Method getConc.

Model methods in R package chemosensors.
}

