% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemosensorsClass.R, R/ConcNoiseModelClass.R, R/DriftNoiseModelClass.R, R/SensorModelClass.R, R/SensorNoiseModelClass.R, R/SensorArrayClass.R, R/methods.R
\docType{methods}
\name{noise-methods}
\alias{csd,ConcNoiseModel-method}
\alias{csd<-}
\alias{csd<-,ConcNoiseModel-method}
\alias{dsd,DriftNoiseModel-method}
\alias{dsd<-}
\alias{dsd<-,DriftNoiseModel-method}
\alias{noise-methods}
\alias{nsd<-}
\alias{nsd<-,ConcNoiseModel-method}
\alias{nsd<-,SensorArray-method}
\alias{nsd<-,SensorModel-method}
\alias{nsd<-,SensorNoiseModel-method}
\alias{ssd,SensorNoiseModel-method}
\alias{ssd<-}
\alias{ssd<-,SensorNoiseModel-method}
\title{Method csd.}
\usage{
csd(x)

csd(object) <- value

ssd(x)

ssd(object) <- value

dsd(x)

dsd(object) <- value

nsd(object) <- value

\S4method{csd}{ConcNoiseModel}(x)

\S4method{csd}{ConcNoiseModel}(object) <- value

\S4method{nsd}{ConcNoiseModel}(object) <- value

\S4method{dsd}{DriftNoiseModel}(x)

\S4method{dsd}{DriftNoiseModel}(object) <- value

\S4method{nsd}{SensorModel}(object) <- value

\S4method{ssd}{SensorNoiseModel}(x)

\S4method{ssd}{SensorNoiseModel}(object) <- value

\S4method{nsd}{SensorNoiseModel}(object) <- value

\S4method{nsd}{SensorArray}(object) <- value
}
\description{
Method csd.

Method csd<-.

Method ssd.

Method ssd<-.

Method dsd.

Method dsd<-.

Method nsd<-.

Methods to control the noise models.
}

