% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemosensorsClass.R, R/ChemosensorsClassMethods.R, R/ConcNoiseModelClassMethods.R, R/DriftNoiseModelClassMethods.R, R/ScenarioClassMethods.R, R/SensorArrayClassMethods.R, R/SensorDynamicsClassMethods.R, R/SensorModelClassMethods.R, R/SensorNoiseModelClassMethods.R, R/SorptionModelClassMethods.R, R/methods.R, R/plot.lib.R
\docType{methods}
\name{plot-methods}
\alias{ccol,ANY-method}
\alias{gcol,ANY-method}
\alias{mapcol,ANY-method}
\alias{plot,ConcNoiseModel-method}
\alias{plot,DriftNoiseModel-method}
\alias{plot,Scenario-method}
\alias{plot,SensorArray-method}
\alias{plot,SensorDynamics-method}
\alias{plot,SensorModel-method}
\alias{plot,SensorNoiseModel-method}
\alias{plot,SorptionModel-method}
\alias{plot-methods}
\alias{plotAffinity,ANY-method}
\alias{plotAffinityMap,ANY-method}
\alias{plotAffinitySpace,ANY-method}
\alias{plotBox,ANY-method}
\alias{plotMixture,ANY-method}
\alias{plotPCA,ANY-method}
\alias{plotPolar,ANY-method}
\alias{plotPolarGases,ANY-method}
\alias{plotResponse,ANY-method}
\alias{plotSignal,ANY-method}
\alias{plotTimeline,ANY-method}
\alias{scol,ANY-method}
\title{Method plot.}
\usage{
plot(x, y, ...)

plotPolar(x, y, ...)

plotPCA(x, y, conc, sdata, set, scenario, feature = "transient", air = TRUE,
  mod, center = TRUE, scale = TRUE, pc = 1:2, ...)

plotBox(x, y, conc, sdata, set, scenario, ...)

plotResponse(x, y, ...)

plotMixture(x, y, ...)

plotSignal(x, y, ...)

plotPolarGases(x, y, ...)

plotTimeline(x, y, ...)

plotAffinity(x, y, ...)

plotAffinityMap(x, y, ...)

plotAffinitySpace(x, y, conc, sdata, ...)

ccol(object, ...)

gcol(object, ...)

scol(object, ...)

mapcol(object, ...)

\S4method{plotPolar}{ANY}(x, y, polar = TRUE, geom = "line",
  main = "Model Response: Polar plot", xlab = "Sensor",
  ylab = "Sensor Signal", xlim = NULL, ylim = NULL, graphics = "ggplot",
  ret = TRUE, ...)

\S4method{plotPolarGases}{ANY}(x, y, polar = TRUE, geom = "line",
  main = "Model Response: Polar plot per analyte", xlab = "Sensor",
  ylab = "Sensor Signal", xlim = NULL, ylim = NULL, graphics = "ggplot",
  ret = FALSE, ...)

\S4method{plotTimeline}{ANY}(x, y, conc, sdata, concUnits = "default",
  leg = "none", lwd = 2, lty = 1, col, bty = "n",
  main = "Model Response: Timeline", xlab, ylab, ...)

\S4method{plotAffinitySpace}{ANY}(x, y, type = "points", gases = 1:2,
  main = "Affinity Space", xlab, ylab, graphics = "ggplot",
  geom = "point", ret = FALSE, ...)

\S4method{plotAffinityMap}{ANY}(x, y, gases = 0, main = "Affinity Map",
  xlab, ylab, ...)

\S4method{plotAffinity}{ANY}(x, y, gases = 0, main = "Affinity Map", xlab,
  ylab, ...)

\S4method{plotMixture}{ANY}(x, y, n = 20, concUnits = "default",
  gases = 0, las2 = 1, axes = TRUE, col, nlevels = 10,
  main = "Response to Binary Mixture", xlab, ylab, ...)

\S4method{plotSignal}{ANY}(x, y, conc, set, concUnits = "default",
  main = "Model Response to Concentration", ret = TRUE, ...)

\S4method{ccol}{ANY}(object, conc, pal, ...)

\S4method{gcol}{ANY}(object, gases = 0, pal, ...)

\S4method{mapcol}{ANY}(object, col, n = 100, ...)

\S4method{scol}{ANY}(object, n, start = 0.2, end = 0.8, gamma = 2.2)

\S4method{plot}{ConcNoiseModel}(x, y, ...)

\S4method{plot}{DriftNoiseModel}(x, y, ...)

\S4method{plot}{Scenario}(x, y, ret = TRUE, ...)

\S4method{plot}{SensorArray}(x, y, ...)

\S4method{plot}{SensorDynamics}(x, y, ...)

\S4method{plot}{SensorModel}(x, y, ...)

\S4method{plot}{SensorNoiseModel}(x, y, ...)

\S4method{plot}{SorptionModel}(x, y, ...)

\S4method{plotBox}{ANY}(x, y, conc, sdata, set, scenario, cf, sf, sensors = 1,
  feature = "step", air = TRUE, concUnits = "default", main = "Boxplot",
  xlab, ylab, scales = c("fixed", "free_y"), gcol = FALSE,
  sensor.names = "long", ret = TRUE, ...)

\S4method{plotResponse}{ANY}(x, y, idx = 1, sensor, gas,
  concUnits = "default", n = 100, main = "Model Response", ret = TRUE,
  ...)

\S4method{plotPCA}{ANY}(x, y, conc, sdata, set, scenario,
  feature = "transient", air = TRUE, mod, center = TRUE, scale = TRUE,
  pc = 1:2, concUnits = "default", main = "PCA scoreplot", xlab, ylab,
  size = NULL, alpha = NULL, ret = TRUE, ...)
}
\description{
Method plot.

Method plotPolar.

Method plotPCA.

Method plotBox.

Method plotResponse.

Method plotMixture.

Method plotSignal.

Method plotPolarGases.

Method plotTimeline.

Method plotAffinity.

Method plotAffinityMap.

Method plotAffinitySpace.

Method ccol.

Method gcol.

Method scol.

Method mapcol.

Plot methods in R package chemosensors.
}

