\name{sdata2feature}
\alias{sdata2feature}
\title{Method sdata2feature}
\description{
  Method sdata2feature converts a matrix of sensor data
  into a data frame of features.
}
\details{
  The input parameters are an object, e.g. of class
  \code{SensorArray}, a concentration matrix, a matrix of
  sensor data, and (optionally) a data frame derived from
  conccentraion matrix.
}
