\name{SensorArray}
\alias{SensorArray}
\title{Class SensorArray.}
\description{
  Class \code{\link{SensorArray}} predicts a sensor array
  signal in response to an input concentration matrix by
  passing the data through a list of models.
}
\details{
  The models aggregated into the class are
  {\link{ConcNoiseModel}}, {\link{SensorNoiseModel}} and
  {\link{SensorArrayModel}}.
}
\keyword{SensorArray-class}

