\name{SensorDataModel-class}
\alias{SensorDataModel}
\alias{SensorDataModel-class}
\title{Class SensorDataModel.}
\usage{
  SensorDataModel(model, ...)
}
\arguments{
  \item{model}{Name of the data model. Supported names:
  \code{"pls"}, \code{"pls"}, \code{"mvr"},
  \code{"broken-stick"} and \code{"ispline"}.}

  \item{...}{Parameters of constructor.}
}
\value{
  An object of class \code{SensorDataModel}.
}
\description{
  Class SensorDataModel.

  Constructor method of SensorDataModel Class.
}

