% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemosensorsClass.R, R/ChemosensorsClassMethods.R, R/SensorDynamicsClass.R, R/methods.R
\docType{methods}
\name{enableDyn<-}
\alias{concUnits<-}
\alias{concUnits<-,ANY-method}
\alias{enableDyn<-}
\alias{enableDyn<-,SensorDynamics-method}
\alias{set-methods}
\title{Method enableDyn<-.}
\usage{
enableDyn(object) <- value

concUnits(object) <- value

\S4method{concUnits}{ANY}(object) <- value

\S4method{enableDyn}{SensorDynamics}(object) <- value
}
\description{
Method enableDyn<-.

Method concUnits<-.

Set Methods in R package chemosensors.
}

