\name{get-methods}
\alias{alpha}
\alias{alpha,SorptionModel-method}
\alias{beta}
\alias{beta,SensorModel-method}
\alias{coef}
\alias{coef,ANY-method}
\alias{coefficients}
\alias{coefficients,SensorArray-method}
\alias{coefficients,SensorDynamics-method}
\alias{coefficients,SensorModel-method}
\alias{coefnames}
\alias{coefnames,SensorNoiseModel-method}
\alias{concUnits}
\alias{concUnits,ANY-method}
\alias{concUnitsInt}
\alias{concUnitsInt,ANY-method}
\alias{enableDyn}
\alias{enableDyn,SensorDynamics-method}
\alias{enableSorption}
\alias{enableSorption,SensorArray-method}
\alias{gases}
\alias{gases,ANY-method}
\alias{gases,missing-method}
\alias{get-methods}
\alias{gind}
\alias{gind,ANY-method}
\alias{gind,missing-method}
\alias{gnames}
\alias{gnames,ANY-method}
\alias{gnames,missing-method}
\alias{idx}
\alias{idx,ANY-method}
\alias{modelName}
\alias{modelName,DriftNoiseModel-method}
\alias{modelName,SensorModel-method}
\alias{ncoef}
\alias{ncoef,SensorDynamics-method}
\alias{ncoef,SensorModel-method}
\alias{ncoef,SensorNoiseModel-method}
\alias{ngases}
\alias{ngases,ANY-method}
\alias{ngases,missing-method}
\alias{nsensors}
\alias{nsensors,ANY-method}
\alias{nsensors,SorptionModel-method}
\alias{num}
\alias{num,ANY-method}
\alias{snames}
\alias{snames,ANY-method}
\alias{tunit}
\alias{tunit,Scenario-method}
\alias{tunit,SensorDynamics-method}
\alias{type}
\alias{type,ConcNoiseModel-method}
\alias{type,DriftNoiseModel-method}
\alias{type,SensorNoiseModel-method}
\title{Get Methods in R package chemosensors.}
\description{
  Get Methods in R package chemosensors.

  Method alpha.

  Method beta.

  Method modelName.

  Method tunit.

  Method enableSorption.

  Method enableDyn.

  Method num.

  Method idx.

  Method gases.

  Method gind.

  Method ngases.

  Method gnames.

  Method coefficients.

  Method coef.

  Method nsensors.

  Method snames.

  Method concUnits.

  Method concUnitsInt.

  Method type.

  Method coefnames.

  Method ncoef.
}

