\name{model-methods}
\alias{concMax}
\alias{concMax,ANY-method}
\alias{concMin}
\alias{concMin,ANY-method}
\alias{concSample}
\alias{concSample,ANY-method}
\alias{concSample,Scenario-method}
\alias{concSampleDyn}
\alias{concSampleDyn,ANY-method}
\alias{getConc}
\alias{getConc,Scenario-method}
\alias{model-methods}
\alias{predict}
\alias{predict,ConcNoiseModel-method}
\alias{predict,DriftNoiseModel-method}
\alias{predict,SensorArray-method}
\alias{predict,SensorDynamics-method}
\alias{predict,SensorModel-method}
\alias{predict,SensorNoiseModel-method}
\alias{predict,SorptionModel-method}
\title{Model methods in R package chemosensors.}
\description{
  Model methods in R package chemosensors.

  Method concMin.

  Method concMax.

  Method concSample.

  Method concSampleDyn.

  Method predict.

  Method getConc.
}

