\name{noise-methods}
\alias{csd}
\alias{csd,ConcNoiseModel-method}
\alias{csd<-}
\alias{csd<-,ConcNoiseModel-method}
\alias{dsd}
\alias{dsd,DriftNoiseModel-method}
\alias{dsd<-}
\alias{dsd<-,DriftNoiseModel-method}
\alias{noise-methods}
\alias{nsd<-}
\alias{nsd<-,ConcNoiseModel-method}
\alias{nsd<-,SensorArray-method}
\alias{nsd<-,SensorModel-method}
\alias{nsd<-,SensorNoiseModel-method}
\alias{ssd}
\alias{ssd,SensorNoiseModel-method}
\alias{ssd<-}
\alias{ssd<-,SensorNoiseModel-method}
\title{Methods to control the noise models.}
\description{
  Methods to control the noise models.

  Method csd.

  Method csd<-.

  Method ssd.

  Method ssd<-.

  Method dsd.

  Method dsd<-.

  Method nsd<-.
}

