\name{plot-methods}
\alias{ccol}
\alias{ccol,ANY-method}
\alias{gcol}
\alias{gcol,ANY-method}
\alias{mapcol}
\alias{mapcol,ANY-method}
\alias{plot}
\alias{plot,ConcNoiseModel-method}
\alias{plot,DriftNoiseModel-method}
\alias{plot,Scenario-method}
\alias{plot,SensorArray-method}
\alias{plot,SensorDynamics-method}
\alias{plot,SensorModel-method}
\alias{plot,SensorNoiseModel-method}
\alias{plot,SorptionModel-method}
\alias{plot-methods}
\alias{plotAffinity}
\alias{plotAffinity,ANY-method}
\alias{plotAffinityMap}
\alias{plotAffinityMap,ANY-method}
\alias{plotAffinitySpace}
\alias{plotAffinitySpace,ANY-method}
\alias{plotBox}
\alias{plotBox,ANY-method}
\alias{plotMixture}
\alias{plotMixture,ANY-method}
\alias{plotPCA}
\alias{plotPCA,ANY-method}
\alias{plotPolar}
\alias{plotPolar,ANY-method}
\alias{plotPolarGases}
\alias{plotPolarGases,ANY-method}
\alias{plotResponse}
\alias{plotResponse,ANY-method}
\alias{plotSignal}
\alias{plotSignal,ANY-method}
\alias{plotTimeline}
\alias{plotTimeline,ANY-method}
\alias{scol}
\alias{scol,ANY-method}
\title{Plot methods in R package chemosensors.}
\description{
  Plot methods in R package chemosensors.

  Method plot.

  Method plotPolar.

  Method plotPCA.

  Method plotBox.

  Method plotResponse.

  Method plotMixture.

  Method plotSignal.

  Method plotPolarGases.

  Method plotTimeline.

  Method plotAffinity.

  Method plotAffinityMap.

  Method plotAffinitySpace.

  Method ccol.

  Method gcol.

  Method scol.

  Method mapcol.
}

