\name{set-methods}
\alias{concUnits<-}
\alias{concUnits<-,ANY-method}
\alias{enableDyn<-}
\alias{enableDyn<-,SensorDynamics-method}
\alias{set-methods}
\title{Set Methods in R package chemosensors.}
\description{
  Set Methods in R package chemosensors.

  Method enableDyn<-.

  Method concUnits<-.
}

